[Feature] 대시보드

+++ src/main/java/com/sys/stm/domains/meeting/service/MeetingServiceImpl.java
@@ -3,19 +3,16 @@ package com.sys.stm.domains.meeting.service;
 import com.sys.stm.domains.meeting.dao.MeetingRepository;
 import com.sys.stm.domains.meeting.domain.Meeting;
 import com.sys.stm.domains.meeting.domain.MeetingParticipant;
-import com.sys.stm.domains.meeting.domain.Participant;
 import com.sys.stm.domains.meeting.dto.request.MeetingCreateRequestDTO;
+import com.sys.stm.domains.meeting.dto.request.MeetingUpdateRequestDTO;
 import com.sys.stm.domains.meeting.dto.response.*;
 import com.sys.stm.global.exception.ExceptionMessage;
+import com.sys.stm.global.exception.ForbiddenException;
 import com.sys.stm.global.exception.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.io.UnsupportedEncodingException;
-import java.net.URLEncoder;
-import java.sql.Timestamp;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -97,6 +94,7 @@ public class MeetingServiceImpl implements MeetingService {
         return response;
     }
 
+
     @Transactional(readOnly = true)
     @Override
     public MeetingListPageResponseDTO<MeetingListResponseDTO> getMeetingList(Long projectId, int page, int size, String progressDate, String keyword) {
@@ -174,23 +172,59 @@ public class MeetingServiceImpl implements MeetingService {
     }
 
     @Override
-    public void deleteMeeting(Long meetingId) {
+    public void deleteMeeting(Long meetingId, Long memberId) {
 
         Meeting meeting = validateMeetingExists(meetingId);
 
-        // 회의록 삭제
-        meetingRepository.deleteMeeting(meeting);
+        if(meeting.getMemberId().equals(memberId)) {
+            // 회의록 삭제
+            meetingRepository.deleteMeeting(meeting);
 
-        // 회의록에 연관된 데이터 삭제
-        // 1. 참석자 데이터 삭제
-        meeting.getParticipants().stream()
-                .forEach(
-                        meetingParticipant -> {
-                            meetingParticipantService.deleteParticipant(meetingParticipant.getId());
-                        }
-                );
+            // 회의록에 연관된 데이터 삭제
+            // 1. 참석자 데이터 삭제
+            meeting.getParticipants().stream()
+                    .forEach(
+                            meetingParticipant -> {
+                                meetingParticipantService.deleteParticipant(meetingParticipant.getId());
+                            }
+                    );
+        }else{
+            throw new ForbiddenException(ExceptionMessage.ACCESS_DENIED);
+        }
+
+    }
 
 
+    @Override
+    public void updateMeeting(MeetingUpdateRequestDTO meetingUpdateRequestDTO, Long meetingId,  Long memberId) {
+        Meeting meeting = validateMeetingExists(meetingId).update(meetingUpdateRequestDTO);
+
+        if(meeting.getMemberId().equals(memberId)) {
+            // 회의록에 연관된 데이터 삭제
+            // 참석자 데이터 삭제
+            meeting.getParticipants().stream()
+                    .forEach(
+                            meetingParticipant -> {
+                                meetingParticipantService.deleteParticipant(meetingParticipant.getId());
+                            }
+                    );
+
+            // 생성된 회의록 ID 기준으로 참여자 Entity 리스트 생성
+            for (Long participantMemberId : meetingUpdateRequestDTO.getParticipantIds()) {
+                MeetingParticipant participant = MeetingParticipant.builder()
+                        .meetingId(meeting.getId())  // 생성된 Meeting ID 사용
+                        .memberId(participantMemberId)
+                        .build();
+
+                // 참여자 테이블 저장
+                meetingParticipantService.createParticipant(participant);
+            }
+
+            meetingRepository.createMeeting(meeting);
+
+        }else{
+            throw new ForbiddenException(ExceptionMessage.ACCESS_DENIED);
+        }
     }
 
 
@@ -228,14 +262,6 @@ public class MeetingServiceImpl implements MeetingService {
 
         return result;
     }
-
-
-    @Override
-    public Long createMeetingWithParticipants(MeetingCreateRequestDTO request, Long projectId, Long memberId, List<Long> participantMemberIds) {
-        return 0L;
-    }
-
-
 
 
     protected Meeting validateMeetingExists(Long meetingId) {
